added terraform controller module
Terraform controller is a class that is a layer on top of terraform. It
will create the configuration file, create and destroy infrastructure,
get resources information...

On this commit, the first draft of the controller was created.

We have to create the terraform files dynamically because terraform does
not provide a way to create multiple providers with a `for_each` or
`count` way. Also, the provider of a resource can't specified with a
variables.