CI: add image builder tests in CIV ci pipeline
CIV tool is being executed in IB CI, but it's running a specific release,
so there's a lot of manual work to keep this release updated. If we run
this tests in out own CI and publish the container image, IB can always
get the last version of CIV without worrying if it's going to cause
errors.