Fixed dependencies issues with PyYAML and pytest. In the past days there was a bug affecting pytest. This was reported and tracked here:
* https://github.com/yaml/pyyaml/issues/724

The proposed solution was to downgrade PyYAML to 5.3.1 until it was resolved. Now this issue is fixed in PyYAML, however there are still no updates from pytest, which is currently in v7.4.0 (last update in June 23 2023).

There were isues also encountered in awscli pip package which does not allow newer versions of PyYAML.

Until they update this in next release, we will force PyYAML 5.3.1 to be installed before pytest in our project dependencies.

In this PR we also take the opportunity to update pytest-testinfra to benefit from stability and performance improvements.