AzureConfigBuilderV2: gracefully handle dot in release versions
An osbuild-composer PR#3887 [1] introduces "dot-notation" for distro
versions, meaning that a dot will be used to separate major and minor
version in distro names. For backward compatibility, distro names
without a dot will be still accepted for distros which used to accept
it.

The CIV code parsing a VHD name was not expecting the version to contain
any dot. This results in failures such as [2].

Modify the regular expression to gracefully handle potential dot in the
version string.

[1] https://github.com/osbuild/osbuild-composer/pull/3887
[2] https://gitlab.com/redhat/services/products/image-builder/ci/osbuild-composer/-/jobs/5943494707#L3695

Signed-off-by: Tomáš Hozza <thozza@redhat.com>