Added new files to the otel testing
fixtures.py - There are many fixtures related to this test, In my opinion it is better arranged so in this case.
This fixture file is under otel_packge folder which are fixtures that are spesific only for this testing pakckage.

helpers.py - The helpers file include a method that might be used for many other tests and not be limited by a fixture scope.
This file may include some other generic helpful methods in our tests.